# webapp/src/main/java/ar/edu/itba/paw/webapp/restcontrollers/openapi.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: RankUp REST API
  description: |
    REST API for RankUp.

    Current pagination contract: paginated endpoints accept a zero-based `page`
    query parameter. When pagination links are emitted, they are returned in the
    standard `Link` response header with `self`, `first`, `last`, and, when
    applicable, `prev` and `next` relations.

    Current error contract: known API errors return an `ErrorDTO` JSON body with
    `status`, `error`, and `message`. Validation errors may also include a
    `details` object keyed by field or parameter name. Authentication failures
    return 401 JSON and authorization failures return 403 JSON. Jersey-generated
    media type errors such as 406 and 415 use the same schema.

    HATEOAS links that currently point to resources or filters without a REST
    controller are not documented as available endpoints here. Known pending
    resources/filters include user-created/user-participated tournament lists
    and user-participated match lists.
servers:
  - url: /api
...
paths:
...
  /tournaments:
    get:
      tags: [tournaments]
      summary: List tournaments
      description: Results are paginated with zero-based `page`.
      parameters:
        - in: query
          name: name
          schema:
            type: string
        - in: query
          name: gameId
          schema:
            type: integer
            format: int64
        - in: query
          name: region
          schema:
            $ref: '#/components/schemas/Region'
        - in: query
          name: elo
          schema:
            $ref: '#/components/schemas/Elo'
        - in: query
          name: startDate
          schema:
            type: string
            format: date
        - in: query
          name: endDate
          schema:
            type: string
            format: date
        - in: query
          name: format
          schema:
            type: string
        - in: query
          name: structure
          schema:
            $ref: '#/components/schemas/Structure'
        - in: query
          name: genre
          schema:
            $ref: '#/components/schemas/Genre'
        - in: query
          name: playersPerTeam
          schema:
            type: integer
            minimum: 1
        - $ref: '#/components/parameters/Page'
      responses:
        '200':
          description: Tournament list
          headers:
            Link:
              $ref: '#/components/headers/PaginationLink'
          content:
            application/vnd.tournament.list.v1+json:
              schema:
                $ref: '#/components/schemas/TournamentList'
        '400':
          $ref: '#/components/responses/BadRequestJson'
    post:
      tags: [tournaments]
      summary: Create tournament
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/vnd.tournament.create.v1+json:
            schema:
              $ref: '#/components/schemas/CreateTournamentRequest'
      responses:
        '201':
          description: Tournament created
          headers:
            Location:
              $ref: '#/components/headers/Location'
          content:
            application/vnd.tournament.v1+json:
              schema:
                $ref: '#/components/schemas/Tournament'
        '400':
          $ref: '#/components/responses/BadRequestJson'
        '401':
          $ref: '#/components/responses/UnauthorizedJson'
        '403':
          $ref: '#/components/responses/ForbiddenJson'
        '404':
          $ref: '#/components/responses/NotFoundJson'
        '409':
          $ref: '#/components/responses/ConflictJson'
  /tournaments/{id}:
    get:
      tags: [tournaments]
      summary: Get tournament
      parameters:
        - $ref: '#/components/parameters/TournamentId'
      responses:
        '200':
          description: Tournament
          content:
            application/vnd.tournament.v1+json:
              schema:
                $ref: '#/components/schemas/Tournament'
        '400':
          $ref: '#/components/responses/BadRequestJson'
        '404':
          $ref: '#/components/responses/NotFoundJson'
    put:
      tags: [tournaments]
      summary: Update tournament
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TournamentId'
      requestBody:
        required: true
        content:
          application/vnd.tournament.update.v1+json:
            schema:
              $ref: '#/components/schemas/UpdateTournamentRequest'
      responses:
        '200':
          description: Updated tournament
          content:
            application/vnd.tournament.v1+json:
              schema:
                $ref: '#/components/schemas/Tournament'
        '400':
          $ref: '#/components/responses/BadRequestJson'
        '404':
          $ref: '#/components/responses/NotFoundJson'
        '401':
          $ref: '#/components/responses/UnauthorizedJson'
        '403':
          $ref: '#/components/responses/ForbiddenJson'
  /tournaments/{id}/participants:
    get:
      tags: [tournaments]
      summary: Get tournament participants
      description: Not paginated in the current controller.
      parameters:
        - $ref: '#/components/parameters/TournamentId'
        - in: query
          name: teamSize
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Participant list
          content:
            application/vnd.participant.list.v1+json:
              schema:
                $ref: '#/components/schemas/ParticipantList'
        '400':
          $ref: '#/components/responses/BadRequestJson'
  /tournaments/{id}/rules:
    get:
      tags: [tournaments]
      summary: Get tournament rules PDF
      parameters:
        - $ref: '#/components/parameters/TournamentId'
      responses:
        '200':
          description: Tournament rules PDF
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/BadRequestJson'
        '404':
          $ref: '#/components/responses/NotFoundJson'
  /tournaments/{id}/participants/users:
    post:
      tags: [tournaments]
      summary: Join tournament as user
      description: Joins the authenticated Bearer-token user; no user id is accepted in the request body.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TournamentId'
      responses:
        '204':
          description: Joined
        '400':
          $ref: '#/components/responses/BadRequestJson'
        '401':
          $ref: '#/components/responses/UnauthorizedJson'
        '403':
          $ref: '#/components/responses/ForbiddenJson'
        '404':
          $ref: '#/components/responses/NotFoundJson'
        '409':
          $ref: '#/components/responses/ConflictJson'
  /tournaments/{id}/participants/teams:
    post:
      tags: [tournaments]
      summary: Join tournament as team
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TournamentId'
      requestBody:
        required: true
        content:
          application/vnd.tournament.join-team.v1+json:
            schema:
              $ref: '#/components/schemas/JoinTournamentTeamRequest'
      responses:
        '204':
          description: Joined
        '400':
          $ref: '#/components/responses/BadRequestJson'
        '401':
          $ref: '#/components/responses/UnauthorizedJson'
        '403':
          $ref: '#/components/responses/ForbiddenJson'
        '404':
          $ref: '#/components/responses/NotFoundJson'
  /tournaments/{id}/participants/me:
    delete:
      tags: [tournaments]
      summary: Leave tournament as current user
      description: Removes the authenticated Bearer-token user from the tournament without accepting a user id from the client.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TournamentId'
      responses:
        '204':
          description: Left
        '400':
          $ref: '#/components/responses/BadRequestJson'
        '401':
          $ref: '#/components/responses/UnauthorizedJson'
        '403':
          $ref: '#/components/responses/ForbiddenJson'
        '404':
          $ref: '#/components/responses/NotFoundJson'
  /tournaments/{id}/participants/{participantId}:
    get:
      tags: [tournaments]
      summary: Get tournament participant
      parameters:
        - $ref: '#/components/parameters/TournamentId'
        - $ref: '#/components/parameters/ParticipantId'
      responses:
        '200':
          description: Participant
          content:
            application/vnd.participant.v1+json:
              schema:
                $ref: '#/components/schemas/Participant'
        '400':
          $ref: '#/components/responses/BadRequestJson'
        '404':
          $ref: '#/components/responses/NotFoundJson'
    delete:
      tags: [tournaments]
      summary: Remove participant
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TournamentId'
        - $ref: '#/components/parameters/ParticipantId'
      responses:
        '204':
          description: Removed
        '400':
          $ref: '#/components/responses/BadRequestJson'
        '401':
          $ref: '#/components/responses/UnauthorizedJson'
        '403':
          $ref: '#/components/responses/ForbiddenJson'
        '404':
          $ref: '#/components/responses/NotFoundJson'
  /tournaments/{id}/participants/users/{userId}:
    delete:
      tags: [tournaments]
      summary: Leave tournament as user by id
      description: Legacy id-based endpoint retained for compatibility; SPA clients should use `/tournaments/{id}/participants/me`.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TournamentId'
        - $ref: '#/components/parameters/UserIdPath'
      responses:
        '204':
          description: Left
        '400':
          $ref: '#/components/responses/BadRequestJson'
        '404':
          $ref: '#/components/responses/NotFoundJson'
        '401':
          $ref: '#/components/responses/UnauthorizedJson'
        '403':
          $ref: '#/components/responses/ForbiddenJson'
  /tournaments/{id}/matches:
    get:
      tags: [tournaments]
      summary: Get tournament matches grouped by stage
      description: Not paginated in the current controller.
      parameters:
        - $ref: '#/components/parameters/TournamentId'
        - in: query
          name: group
          schema:
            type: integer
      responses:
        '200':
          description: Match stages
          content:
            application/vnd.matchstagelist.v1+json:
              schema:
                $ref: '#/components/schemas/MatchStageList'
        '400':
          $ref: '#/components/responses/BadRequestJson'
  /tournaments/{id}/matches/{matchId}:
    get:
      tags: [tournaments]
      summary: Get tournament match
      parameters:
        - $ref: '#/components/parameters/TournamentId'
        - $ref: '#/components/parameters/MatchId'
      responses:
        '200':
          description: Match
          content:
            application/vnd.match.v1+json:
              schema:
                $ref: '#/components/schemas/Match'
        '400':
          $ref: '#/components/responses/BadRequestJson'
        '404':
          $ref: '#/components/responses/NotFoundJson'
  /tournaments/{id}/matches/{matchId}/results:
    put:
      tags: [tournaments]
      summary: Set match results
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TournamentId'
        - $ref: '#/components/parameters/MatchId'
      requestBody:
        required: true
        content:
          application/vnd.match.results.v1+json:
            schema:
              $ref: '#/components/schemas/SetMatchResultsRequest'
      responses:
        '204':
          description: Updated
        '400':
          $ref: '#/components/responses/BadRequestJson'
        '401':
          $ref: '#/components/responses/UnauthorizedJson'
        '403':
          $ref: '#/components/responses/ForbiddenJson'
        '404':
          $ref: '#/components/responses/NotFoundJson'
        '409':
          $ref: '#/components/responses/ConflictJson'
  /tournaments/{id}/status:
    put:
      tags: [tournaments]
      summary: Update tournament status
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TournamentId'
      requestBody:
        required: true
        content:
          application/vnd.tournament.status.v1+json:
            schema:
              $ref: '#/components/schemas/TournamentStatusRequest'
      responses:
        '200':
          description: Updated tournament status
          content:
            application/vnd.tournament.v1+json:
              schema:
                $ref: '#/components/schemas/Tournament'
        '400':
          $ref: '#/components/responses/BadRequestJson'
        '404':
          $ref: '#/components/responses/NotFoundJson'
        '401':
          $ref: '#/components/responses/UnauthorizedJson'
        '403':
          $ref: '#/components/responses/ForbiddenJson'
        '409':
          $ref: '#/components/responses/ConflictJson'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  headers:
...
    Location:
      description: Absolute URI of the created resource/token.
      schema:
        type: string
        format: uri
    PaginationLink:
      description: RFC 8288 pagination links with `self`, `first`, `last`, and optional `prev`/`next`.
      schema:
        type: string
  parameters:
    Page:
      in: query
      name: page
      schema:
        type: integer
        minimum: 0
        default: 0
      description: Zero-based page index.
...
    TournamentId:
      in: path
      name: id
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
...
  responses:
    BadRequestJson:
      description: Invalid request syntax, parameters, DTO validation, or business input.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFoundJson:
      description: Requested resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ConflictJson:
      description: Request conflicts with the current resource state.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    UnauthorizedJson:
      description: Missing, empty, or invalid Bearer token.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ForbiddenJson:
      description: Authenticated user does not have the required permission.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Region:
      type: string
      enum: [NA, LAS, LAN, BR, EUW, EUNE, OCE, ASIA]
    Elo:
      type: string
      enum: [LOW, MID, HIGH, FREE]
    Structure:
      type: string
      enum: [ELIMINATION, LEAGUE, HYBRID]
    Genre:
      type: string
      enum: [MOBA, FPS, Fighting, TPS, BattleRoyale, RTS, Sports, DGC, MOBILE]
...
    Tournament:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        region:
          type: string
        elo:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        format:
          type: string
        structure:
          type: string
        maxParticipants:
          type: integer
        openInscriptions:
          type: boolean
        finished:
          type: boolean
        groupStage:
          type: boolean
        tournamentStarted:
          type: boolean
        rating:
          type: number
          format: float
        serverName:
          type: string
        discordChannel:
          type: string
        links:
          type: array
          items:
            $ref: '#/components/schemas/Link'
    CreateTournamentRequest:
      type: object
      properties:
        name:
          type: string
        gameId:
          type: integer
          format: int64
          minimum: 1
        region:
          $ref: '#/components/schemas/Region'
        elo:
          $ref: '#/components/schemas/Elo'
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        structure:
          $ref: '#/components/schemas/Structure'
        maxParticipants:
          type: integer
          minimum: 1
        formatId:
          type: integer
          format: int64
          minimum: 1
        imageBase64:
          type: string
        rulesBase64:
          type: string
        serverName:
          type: string
        serverPassword:
          type: string
        discordChannel:
          type: string
      required: [name, gameId, region, elo, startDate, endDate, structure, maxParticipants, formatId]
...
    JoinTournamentTeamRequest:
      type: object
      properties:
        teamId:
          type: integer
          format: int64
          minimum: 1
        members:
          type: array
          minItems: 1
          items:
            type: integer
            format: int64
            minimum: 1
      required: [teamId, members]
...
    TournamentList:
      type: array
      items:
        $ref: '#/components/schemas/Tournament'
    ParticipantList:
      type: array
      items:
        $ref: '#/components/schemas/Participant'
...